Enable more storage tests to work independent of the providerless tag
