+++ test/e2e/storage/flexvolume_online_resize.go
@@ -1,6 +1,3 @@
-//go:build !providerless
-// +build !providerless
-
 /*
 Copyright 2018 The Kubernetes Authors.
