+++ test/e2e/storage/generic_persistent_volume-disruptive.go
@@ -1,6 +1,3 @@
-//go:build !providerless
-// +build !providerless
-
 /*
 Copyright 2018 The Kubernetes Authors.
