+++ test/e2e/storage/nfs_persistent_volume-disruptive.go
@@ -1,6 +1,3 @@
-//go:build !providerless
-// +build !providerless
-
 /*
 Copyright 2016 The Kubernetes Authors.
